+++ src/topk.h
@@ -11,12 +11,12 @@
  * explicitly suffixed entry point (mirroring the gpu_layout_kernels
  * convention): CPU executables call topk_int_omp_cpu, GPU executables call
  * topk_int_omp_gpu. */
-void topk_int_omp_cpu(const int *dist, int64_t N, int64_t M, int K,
-                      int *out_dist, int *out_idx, int dev_id);
-void topk_int_omp_gpu(const int *dist, int64_t N, int64_t M, int K,
-                      int *out_dist, int *out_idx, int dev_id);
+void topk_int_omp_cpu(const uint64_t *dist, size_t N, size_t M, size_t K,
+                      uint64_t *out_dist, size_t *out_idx, int dev_id);
+void topk_int_omp_gpu(const uint64_t *dist, size_t N, size_t M, size_t K,
+                      uint64_t *out_dist, size_t *out_idx, int dev_id);
 
 /* Unified name reserved for the future internal filtered-results wrapper. */
-void topk_int_omp(const int *dist, int64_t N, int64_t M, int K,
-                     int *out_dist, int *out_idx, int dev_id);
+void topk_int_omp(const uint64_t *dist, size_t N, size_t M, size_t K,
+                  uint64_t *out_dist, size_t *out_idx, int dev_id);
