+++ src/topk_gpu.c
@@ -8,8 +8,8 @@
 #ifndef NOGPU
 
 #if defined(__clang__) || defined(__INTEL_LLVM_COMPILER) || defined(__llvm__)
-void topk_int_omp_gpu(const int *dist, int64_t N, int64_t M, int K,
-                      int *out_dist, int *out_idx, int dev_id) {
+void topk_int_omp_gpu(const uint64_t *dist, size_t N, size_t M, size_t K,
+                      uint64_t *out_dist, size_t *out_idx, int dev_id) {
   if (N <= 0 || M <= 0 || K <= 0)
     return;
   if (K > M)
@@ -25,16 +25,16 @@ void topk_int_omp_gpu(const int *dist, int64_t N, int64_t M, int K,
     // work directly in the output buffers without malloc/free.
 #pragma omp target teams distribute num_teams(N) thread_limit(1)               \
     is_device_ptr(dist) device(dev_id)
-    for (int64_t row = 0; row < N; ++row) {
+    for (size_t row = 0; row < N; ++row) {
 
       // Because thread_limit is 1, this executes sequentially by the master
       // thread of the team. No inner parallel region is needed.
-      int *my_heap_d = out_dist + row * K;
-      int *my_heap_i = out_idx + row * K;
-      const int *row_dist = dist + row * M;
+      uint64_t *my_heap_d = out_dist + row * K;
+      size_t *my_heap_i = out_idx + row * K;
+      const uint64_t *row_dist = dist + row * M;
 
       // --- PHASE 1: Initialize the max-heap ---
-      for (int k = 0; k < K; ++k) {
+      for (size_t k = 0; k < K; ++k) {
         my_heap_d[k] = row_dist[k];
         my_heap_i[k] = k;
       }
@@ -43,22 +43,22 @@ void topk_int_omp_gpu(const int *dist, int64_t N, int64_t M, int K,
       heapify(my_heap_d, my_heap_i, K, 0, 1);
 
       // --- PHASE 2: Stream remaining M elements ---
-      for (int64_t j = K; j < M; ++j) {
-        int d = row_dist[j];
+      for (size_t j = K; j < M; ++j) {
+        uint64_t d = row_dist[j];
         if (d < my_heap_d[0]) {
           my_heap_d[0] = d;
-          my_heap_i[0] = (int)j;
+          my_heap_i[0] = j;
           sift_down(my_heap_d, my_heap_i, 0, K, 0, 1);
         }
       }
 
       // --- PHASE 3: Convert max-heap to sorted ascending in-place ---
-      for (int i = K - 1; i > 0; --i) {
-        int td = my_heap_d[0];
+      for (size_t i = K - 1; i > 0; --i) {
+        uint64_t td = my_heap_d[0];
         my_heap_d[0] = my_heap_d[i];
         my_heap_d[i] = td;
 
-        int ti = my_heap_i[0];
+        size_t ti = my_heap_i[0];
         my_heap_i[0] = my_heap_i[i];
         my_heap_i[i] = ti;
 
@@ -69,8 +69,8 @@ void topk_int_omp_gpu(const int *dist, int64_t N, int64_t M, int K,
 }
 
 #else
-void topk_int_omp_gpu(const int *dist, int64_t N, int64_t M, int K,
-                      int *out_dist, int *out_idx, int dev_id) {
+void topk_int_omp_gpu(const uint64_t *dist, size_t N, size_t M, size_t K,
+                      uint64_t *out_dist, size_t *out_idx, int dev_id) {
   if (N <= 0 || M <= 0 || K <= 0)
     return;
   if (K > M)
@@ -85,15 +85,15 @@ void topk_int_omp_gpu(const int *dist, int64_t N, int64_t M, int K,
     // The runtime handles grouping these threads into optimal warps/teams.
 #pragma omp target teams distribute parallel for is_device_ptr(dist)           \
     device(dev_id)
-    for (int64_t row = 0; row < N; ++row) {
+    for (size_t row = 0; row < N; ++row) {
 
       // Each thread works directly in the output buffer
-      int *my_heap_d = out_dist + row * K;
-      int *my_heap_i = out_idx + row * K;
-      const int *row_dist = dist + row * M;
+      uint64_t *my_heap_d = out_dist + row * K;
+      size_t *my_heap_i = out_idx + row * K;
+      const uint64_t *row_dist = dist + row * M;
 
       // --- PHASE 1: Initialize the max-heap ---
-      for (int k = 0; k < K; ++k) {
+      for (size_t k = 0; k < K; ++k) {
         my_heap_d[k] = row_dist[k];
         my_heap_i[k] = k;
       }
@@ -102,22 +102,22 @@ void topk_int_omp_gpu(const int *dist, int64_t N, int64_t M, int K,
       heapify(my_heap_d, my_heap_i, K, 0, 1);
 
       // --- PHASE 2: Stream remaining M elements ---
-      for (int64_t j = K; j < M; ++j) {
-        int d = row_dist[j];
+      for (size_t j = K; j < M; ++j) {
+        uint64_t d = row_dist[j];
         if (d < my_heap_d[0]) {
           my_heap_d[0] = d;
-          my_heap_i[0] = (int)j;
+          my_heap_i[0] = j;
           sift_down(my_heap_d, my_heap_i, 0, K, 0, 1);
         }
       }
 
       // --- PHASE 3: Convert max-heap to sorted ascending in-place ---
-      for (int i = K - 1; i > 0; --i) {
-        int td = my_heap_d[0];
+      for (size_t i = K - 1; i > 0; --i) {
+        uint64_t td = my_heap_d[0];
         my_heap_d[0] = my_heap_d[i];
         my_heap_d[i] = td;
 
-        int ti = my_heap_i[0];
+        size_t ti = my_heap_i[0];
         my_heap_i[0] = my_heap_i[i];
         my_heap_i[i] = ti;
